requirements.txt
Flask
gunicorn
tensorflow
numpy
pillow
pandas
numpy
sentence-transformers
scikit-learn